+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.0
 
+ - Add horizontal animation on each button left padding
  - Make sure to access settings file from anywhere in the project
  - Change background and text color of buttons
  - Change tool configuration from settings window